+++ R/addDatabase.R
@@ -49,6 +49,7 @@ addDatabase <- function(pps, db=NULL, verbose=TRUE) {
                 paste0(db[!isSqlite], collapse = ', '))
         db <- db[isSqlite]
     }
+    db <- normalizePath(db)
     if(verbose) {
         cat(paste0('Adding ', length(db), 'databases:\n  ', printN(basename(db), 6, collapse='\n  '), '\n'))
         # if(length(db) > 6) {