+++ R/loadPamguardXML.R
@@ -25,6 +25,7 @@ loadPamguardXML <- function(x) {
     warning('File ', x, ' does not exist')
     return(NULL)
   }
+  x <- normalizePath(x)
   pgxml <- read_xml(x)
   psf <- xml_find_all(pgxml, 'INFO') %>% 
     xml_attr('CONFIGURATION')